Use external js/css files for spa and update nginx config
Earlier nginx was configured to enable auth_request module for all
uris prefixed with /. This is however a problem if there are other
paths (especially static files) under this uri prefix because then
authentication will happen for all these requests. The fix is to
explicitly disable auth_request module for known paths.